+++ src/dactyl_keyboard/cad/bottom.clj
@@ -228,9 +228,9 @@
 (def holes-in-main-body (fastener-fn 3 (any-type ::main ::centre)))
 (def holes-in-main-plate (fastener-fn 3 (any-type ::main ::centre) true))
 (def holes-in-left-housing-body (fastener-fn 3 (any-type ::centre) false true))
-(def holes-in-left-housing-plate (fastener-fn 3 (any-type ::centre) true true))
 (def holes-in-wrist-body (fastener-fn 3 (any-type ::wrist)))
 (def holes-in-wrist-plate (fastener-fn 3 (any-type ::wrist) true))
+(def lateral-holes (fastener-fn 3 (any-type ::main ::centre ::wrist) true))
 (def inserts (fastener-fn 4))
 
 
@@ -341,13 +341,12 @@
     (to-3d getopt (case-positive-2d getopt))))
 
 (defn case-complete
-  "A printable model of a case bottom plate in one piece."
+  "A printable model of a case’s bottom plate in one piece."
   [getopt]
   (model/color (:bottom-plate colours)
     (maybe/difference
       (case-positive getopt)
-      (holes-in-main-plate getopt)
-      (holes-in-left-housing-plate getopt))))
+      (lateral-holes getopt))))  ; May be overkill.
 
 
 ;;;;;;;;;;;;;;;;;
@@ -373,29 +372,46 @@
       (holes-in-wrist-plate getopt))))
 
 
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; Combined Case and Wrist-Rest Plates ;;
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;;;;;;;;;;;;;;;;;;;;;
+;; Combined Plates ;;
+;;;;;;;;;;;;;;;;;;;;;
 
-(defn combined-positive
-  "A combined bottom plate for case and wrist rest.
-  This assumes the use of a threaded-style wrist rest, but will obviously
-  prevent the use of threaded fasteners in adjusting the wrist rest.
-  This is therefore recommended only where there is no space available between
-  case and wrist rest."
+(defn- lateral-positive
+  "A combined bottom plate for everything but the central housing.
+  Where a wrist rest is included, this function assumes the use of a solid or
+  threaded wrist rest and will attach its bottom plate to the main body’s.
+  Where plates abut or overlap, this will obviously prevent the use of threaded
+  fasteners in adjusting the wrist rest."
   [getopt]
   (model/union
     (wall-base-3d getopt)
     (to-3d getopt
-      (model/union
+      (maybe/union
         (case-positive-2d getopt)
-        (wrist/hulled-block-pairs getopt)
-        (wrist-positive-2d getopt)))))
+        (when (and (getopt :wrist-rest :include)
+                   (getopt :wrist-rest :bottom-plate :include))
+          (wrist/hulled-block-pairs getopt)
+          (wrist-positive-2d getopt))))))
+
+(defn- bilateral
+  "Fit passed shape onto both sides of a central housing, if any."
+  [getopt shape]
+  (maybe/union
+    shape
+    (when (getopt :central-housing :derived :include-main)
+      (model/mirror [-1 0 0] shape))))
+
+(defn combined-positive
+  "The positive space of a combined bottom plate: One plate that extends to
+  cover the central housing, if that’s included, and/or the wrist rests, if
+  they’re included."
+  [getopt]
+  (bilateral getopt (lateral-positive getopt)))
 
 (defn combined-complete
+  "A combined plate with holes."
   [getopt]
   (model/color (:bottom-plate colours)
     (maybe/difference
       (combined-positive getopt)
-      (holes-in-main-plate getopt)
-      (holes-in-wrist-plate getopt))))
+      (bilateral getopt (lateral-holes getopt)))))